% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app3.R
\name{ctxsearch}
\alias{ctxsearch}
\title{ssrch demo with metadata documents from 68 cancer transcriptomics studies}
\usage{
ctxsearch()
}
\value{
Simply starts an app.
}
\description{
ssrch demo with metadata documents from 68 cancer transcriptomics studies
}
\note{
The metadata were derived by extracting sample.attributes
fields from a search with github.com/seandavi/SRAdbV2.  The
sample.attributes content varies between studies and sometimes
between experiments within studies.  The field sets were
unified with the sampleAtts function of github.com/vjcitn/HumanTranscriptomeCompendium.  After unification records were stacked and CSVs were
written.
}
\examples{
if (interactive()) {
  oask = options()$example.ask
  options(example.ask=FALSE)
  try(ctxsearch2())
  options(example.ask=oask)
}
}
