% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twolev.R
\name{DocSet}
\alias{DocSet}
\title{constructor for DocSet}
\usage{
DocSet(kw2docs, docs2recs, docs2kw, titles, doc_retriever)
}
\arguments{
\item{kw2docs}{an environment mapping keywords to documents}

\item{docs2recs}{an environment mapping document identifiers to records}

\item{docs2kw}{an environment mapping documents to keywords}

\item{doc_retriever}{a function that, given a document identifier, will produce the document}
}
\description{
constructor for DocSet
}
