% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docset_cancer68}
\alias{docset_cancer68}
\title{DocSet instance with metadata from 68 cancer studies}
\format{S4 class DocSet defined in ssrch}
\usage{
docset_cancer68
}
\description{
DocSet instance with metadata from 68 cancer studies
}
\keyword{datasets}
