% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ds_can1009b}
\alias{ds_can1009b}
\title{DocSet instance with metadata from 1009 cancer studies}
\format{S4 class DocSet defined in ssrch}
\usage{
ds_can1009b()
}
\value{
DocSet instance
}
\description{
DocSet instance with metadata from 1009 cancer studies
}
\note{
This is part of a sequence of datasets assessing how
far we can go with environments of keywords.  Annotation for
43000 samples is indexed here.
}
\examples{
ds_can1009b()
}
