% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twolev.R
\docType{class}
\name{DocSet-class}
\alias{DocSet-class}
\alias{kw2docs}
\alias{docs2kw}
\alias{docs2recs}
\alias{searchDocs}
\alias{retrieve_doc}
\title{Container for simple documents with arbitrary numbers/shapes of records}
\usage{
kw2docs(sdata)

docs2kw(sdata)

docs2recs(sdata)

searchDocs(string, obj, ...)

retrieve_doc(x, obj, ...)
}
\arguments{
\item{sdata}{instance of srchData class}

\item{string}{character(1) query string}

\item{obj}{instance of DocSet class}
}
\description{
Container for simple documents with arbitrary numbers/shapes of records

utilities for ssrch
}
