% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDoc.R
\name{parseDoc}
\alias{parseDoc}
\title{parse a document and place content in a DocSet}
\usage{
parseDoc(csv, DocSetInstance = NULL,
  rec_id_field = "experiment.accession",
  exclude_fields = c("study.accession"), title = NA_character_,
  cleanFields = list(".*id$", ".name$", "_name$"))
}
\arguments{
\item{csv}{a character(1) CSV file path}

\item{rec_id_field}{character(1) field in CSV identifying records}

\item{exclude_fields}{character vector of fields to ignore while parsing}

\item{title}{character(1) document title}

\item{cleanFields}{list of regular expressions identifying fields to ignre}
}
\description{
parse a document and place content in a DocSet
}
