% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_publ_dates}
\alias{study_publ_dates}
\title{publication dates for 6000 SRA transcriptome studies}
\format{data.frame}
\usage{
study_publ_dates
}
\description{
publication dates for 6000 SRA transcriptome studies
}
\keyword{datasets}
