% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urls68}
\alias{urls68}
\title{pubmed URLs for subset of 68 cancer studies}
\format{character vector}
\usage{
urls68
}
\description{
pubmed URLs for subset of 68 cancer studies
}
\keyword{datasets}
