% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twolev.R
\name{DocSet}
\alias{DocSet}
\title{constructor for DocSet}
\usage{
DocSet(kw2docs = new.env(hash = TRUE), docs2recs = new.env(hash =
  TRUE), docs2kw = new.env(hash = TRUE), titles = character(),
  urls = character(), doc_retriever = function(...) NULL)
}
\arguments{
\item{kw2docs}{an environment mapping keywords to documents}

\item{docs2recs}{an environment mapping document identifiers to records}

\item{docs2kw}{an environment mapping documents to keywords}

\item{titles}{a named character vector with titles; names are document identifiers}

\item{urls}{a named character vector with document-associated URLs; names are document identifiers}

\item{doc_retriever}{a function that, given a document identifier, will produce the document}
}
\value{
instance of DocSet
}
\description{
constructor for DocSet
}
\note{
Titles must be bound in post-hoc.  parseDoc produces
data including parsed titles but does not bind the title into
the resulting object.
}
\examples{
getClass("DocSet")
}
