% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app4.R
\name{docset_searchapp}
\alias{docset_searchapp}
\title{interactive app for ssrch DocSet instances}
\usage{
docset_searchapp(docset)
}
\arguments{
\item{docset}{an instance of DocSet}
}
\value{
Simply starts an app.
}
\description{
interactive app for ssrch DocSet instances
}
\examples{
if (interactive()) {
  oask = options()$example.ask
  options(example.ask=FALSE)
  n1 = try(docset_searchapp(ssrch::docset_cancer68))
  str(n1)
  options(example.ask=oask)
}
}
