% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twolev.R
\docType{class}
\name{DocSet-class}
\alias{DocSet-class}
\alias{kw2docs}
\alias{docs2kw}
\alias{docs2recs}
\alias{searchDocs}
\alias{retrieve_doc}
\title{Container for simple documents with arbitrary numbers/shapes of records}
\usage{
kw2docs(sdata)

docs2kw(sdata)

docs2recs(sdata)

searchDocs(string, obj, ...)

retrieve_doc(x, obj, ...)
}
\arguments{
\item{sdata}{instance of srchData class}

\item{string}{character(1) query string}

\item{obj}{instance of DocSet class}

\item{\dots}{passed to base::grep}

\item{x}{character(1) document identifier}
}
\value{
an environment

an environment

an environment

a data.frame with tokens queried (hits) and associated document ids (docs)

result of calling obj@doc_retriever on arguments x, ...
}
\description{
Container for simple documents with arbitrary numbers/shapes of records

utilities for ssrch
}
\examples{
getClass("DocSet")
}
