% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDoc.R
\name{parseDoc}
\alias{parseDoc}
\title{parse a document and place content in a DocSet}
\usage{
parseDoc(csv, DocSetInstance = NULL,
  rec_id_field = "experiment.accession",
  exclude_fields = c("study.accession"), title = NA_character_,
  cleanFields = list("..*id$", ".name$", "_name$", "checksum", "isolate",
  "filename", "^ID$", "barcode", "Sample.Name"))
}
\arguments{
\item{csv}{a character(1) CSV file path}

\item{DocSetInstance}{if NULL, DocSet is initialized in this
function, otherwise the instance is updated with new content}

\item{rec_id_field}{character(1) field in CSV identifying records}

\item{exclude_fields}{character vector of fields to ignore while parsing}

\item{title}{character(1) document title}

\item{cleanFields}{list of regular expressions identifying fields to ignre}
}
\value{
instance of DocSet
}
\description{
parse a document and place content in a DocSet
}
\examples{
myob = ssrch::docset_cancer68
td = tempdir()
alld = ls(docs2kw(myob))
r1 = retrieve_doc(alld[1], myob)
expo = write.csv(r1, paste0(td, "/expo.csv"))
parseDoc(paste0(td, "/expo.csv"), title=ssrch::titles68[alld[1]])
}
