% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2envs.R
\name{kw2stud}
\alias{kw2stud}
\alias{kw2expt}
\alias{expt2kw}
\alias{stud2kw}
\alias{searchStudies}
\alias{retrieveStudies}
\title{utilities for ssrch}
\usage{
kw2stud(sdata)

kw2expt(sdata)

expt2kw(sdata)

stud2kw(sdata)

searchStudies(string, obj, ...)

retrieveStudies(keys, obj)
}
\arguments{
\item{sdata}{instance of srchData class}
}
\description{
utilities for ssrch
}
